make router a component
