+++ src/app/components/Router/get-location.js
@@ -1,0 +1,7 @@
+const getLocation = Location => ({
+  pathname: Location.pathname,
+  hash: Location.hash,
+  query: Location.query
+});
+
+export default getLocation;
